Fixing compilation issues in eval and Some SAFplus running issues

+++ src/examples/eval/src/app/csa102Comp/clCompAppMain.c
@@ -173,7 +173,7 @@ int main(int argc, char *argv[])
 
     clEoMyEoIocPortGet(&iocPort);
     
-    clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. Initializing\n", appName.length, appName.value, mypid);
+    clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. Initializing\n", appName.length, appName.value, (int ) mypid);
     clprintf (CL_LOG_SEV_INFO, "   IOC Address             : 0x%x\n", clIocLocalAddressGet());
     clprintf (CL_LOG_SEV_INFO, "   IOC Port                : 0x%x\n", iocPort);
 
@@ -222,7 +222,7 @@ int main(int argc, char *argv[])
 errorexit:
 
     clprintf (CL_LOG_SEV_ERROR, "Component [%.*s] : PID [%d]. Initialization error [0x%x]\n",
-              appName.length, appName.value, mypid, rc);
+              appName.length, appName.value, (int) mypid, rc);
 
     return -1;
 }
@@ -239,7 +239,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
     SaAisErrorT rc = SA_AIS_OK;
 
     clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. Terminating\n",
-              compName->length, compName->value, mypid);
+              compName->length, compName->value, (int) mypid);
 
     /*
      * Unregister with AMF and respond to AMF saying whether the
@@ -250,7 +250,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
         goto errorexit;
 
     clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. Terminated\n",
-              compName->length, compName->value, mypid);
+              compName->length, compName->value, (int) mypid);
     saAmfResponse(amfHandle, invocation, SA_AIS_OK);
     clEvalAppLogStreamClose(gEvalLogStream);
 
@@ -261,7 +261,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
 errorexit:
 
     clprintf (CL_LOG_SEV_ERROR, "Component [%.*s] : PID [%d]. Termination error [0x%x]\n",
-              compName->length, compName->value, mypid, rc);
+              compName->length, compName->value, (int) mypid, rc);
 
     return;
 }
@@ -346,7 +346,7 @@ void clCompAppAMFCSISet(SaInvocationT       invocation,
      */
 
     clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. CSI Set Received\n", 
-              compName->length, compName->value, mypid);
+              compName->length, compName->value, (int) mypid);
 
     clCompAppAMFPrintCSI(csiDescriptor, haState);
 
@@ -435,7 +435,7 @@ void clCompAppAMFCSIRemove(SaInvocationT  invocation,
                            SaAmfCSIFlagsT csiFlags)
 {
     clprintf (CL_LOG_SEV_INFO, "Component [%.*s] : PID [%d]. CSI Remove Received\n", 
-              compName->length, compName->value, mypid);
+              compName->length, compName->value, (int) mypid);
 
     clprintf (CL_LOG_SEV_INFO, "   CSI                     : %.*s\n", csiName->length, csiName->value);
     clprintf (CL_LOG_SEV_INFO, "   CSI Flags               : 0x%d\n", csiFlags);
